pyproject.toml
[tool.poetry]
name = "social-media-api"
version = "1.0.0"
description = "FastAPI додаток з MVC архітектурою для соціальної мережі"
authors = ["Developer <dev@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "models" },
    { include = "controllers" },
    { include = "services" },
    { include = "schemas" },
    { include = "utils" },
    { include = "database" },
]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.115.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.0"
pymysql = "^1.1.0"
cryptography = "^41.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.0.0"}
python-dotenv = "^1.0.0"
cachetools = "^5.3.0"

[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
pytest = "^7.4.0"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/username/social-media-api"
"Repository" = "https://github.com/username/social-media-api"
"Documentation" = "http://localhost:8000/docs"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88 